pyproject.toml
[project]
name = "ossi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.9.2",
    "numpy>=2.1.3",
    "pyserial>=3.5",
    "ruff>=0.7.4",
]
